+++ packages/flopper/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @palett/flopper
 
+## 0.12.17
+
+### Patch Changes
+
+- fix dependency issue in @palett/projector
+- Updated dependencies
+  - @palett/pres@0.12.17
+  - @palett/color-algebra@0.12.17
+  - @palett/convert@0.12.17
+  - @palett/munsell@0.12.17
+
 ## 0.12.16
 
 ### Patch Changes