+++ packages/projector/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @palett/projector
 
+## 0.12.17
+
+### Patch Changes
+
+- fix dependency issue in @palett/projector
+- Updated dependencies
+  - @palett/pres@0.12.17
+  - @palett/convert@0.12.17
+  - @palett/dye@0.12.17
+  - @palett/presets@0.12.17
+
 ## 0.12.16
 
 ### Patch Changes